requirements.txt
MySQL-python==1.2.5
psycopg2-binary==2.7.4
argparse==1.2.1
pygeoip==0.3.1
pyparsing==2.0.2
wsgiref==0.1.2
